# src/main/resources/openapi/openapi.yml
openapi: 3.0.3

info:
  title: Authentication Service
  version: 1.0.0

servers:
  - url: http://localhost:8080

paths:
  /auth/signup:
    post:
      operationId: signUp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpRequest'
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignUpResponse'
  /auth/login:
    post:
      operationId: logIn
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogInRequest'
      responses:
        '200':
          description: Ok
          content:
            applicaiton/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
  /tokens/refresh:
    post:
      operationId: refreshTokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokensRequest'
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'

components:
  schemas:
    DefaultNonEmptyString:
      type: string
      minLength: 1
      maxLength: 255
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        username:
          type: string
        password:
          type: string
          writeOnly: true
    TokenPair:
      type: object
      properties:
        access:
          type: string
        refresh:
          type: string
    SignUpRequest:
      type: object
      properties:
        username:
          type: string
          format: email
        password:
          $ref: '#/components/schemas/DefaultNonEmptyString'
    SignUpResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        tokens:
          $ref: '#/components/schemas/TokenPair'
    LogInRequest:
      type: object
      properties:
        username:
          type: string
          format: email
        password:
          $ref: '#/components/schemas/DefaultNonEmptyString'
    RefreshTokensRequest:
      type: object
      properties:
        refreshToken:
          type: string